build: restructure and simplify macro-related crates
The previous approach meant we had a total of 4 crates.  In this new approach,
we put the `derive(Lens)` impl into the `lens-derive` crate, and the impl for
the `lens!` procedural macro into the `lens-macro` crate so that it can work
with proc_macro_hack.
