+++ .github/workflows/build_and_test.yaml
@@ -54,4 +54,4 @@ jobs:
       - name: Test
         uses: actions-rs/cargo@v1
         with:
-          command: test
+          command: test -all
