+++ lens-derive/Cargo.toml
@@ -1,10 +1,10 @@
 [package]
-name = "pl-lens-macros-hack-impl"
+name = "pl-lens-derive"
 version = "1.0.0"
 edition = "2018"
 authors = ["Chris Campbell <campbell@plausible.coop>"]
 license = "MIT"
-description = "Temporary crate used to implement the `lens!` procedural macro."
+description = "Provides procedural `derive` macros that are used in conjuction with the `pl-lens` crate."
 keywords = ["plausible", "lens", "functional"]
 homepage = "https://github.com/plausiblelabs/lens-rs"
 repository = "https://github.com/plausiblelabs/lens-rs"
@@ -15,6 +15,5 @@ proc-macro = true
 
 [dependencies]
 proc-macro2 = "1.0"
-proc-macro-hack = "0.5"
 quote = "1.0"
-syn = { version = "1.0", features = ["full", "extra-traits"] }
+syn = "1.0"
