+++ lens-derive/README.md
@@ -1,0 +1,11 @@
+# pl-lens-macros
+
+Provides the custom `Lenses` derive macro that is used in conjunction with
+the `pl-lens` crate.
+
+This crate is automatically re-exported by the `pl-lens` crate, so you generally
+don't need to work with it directly.
+
+# License
+
+`pl-lens-derive` is distributed under an MIT license.  See LICENSE for more details.
