--- lens-macros-hack-impl/README.md
@@ -1,12 +1,0 @@
-# pl-lens-macros-impl
-
-This is a temporary crate used to implement the `lens!` procedural macro, which requires
-the `proc-macro-hack` crate until such a time that procedural macros are officially
-supported in expression position.
-
-This crate is automatically re-exported by the `pl-lens` crate, so you generally
-don't need to work with it directly.
-
-# License
-
-`pl-lens-macros-impl` is distributed under an MIT license.  See LICENSE for more details.
