--- lens-macros-hack/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "pl-lens-macros-hack"
-version = "1.0.0"
-edition = "2018"
-authors = ["Chris Campbell <campbell@plausible.coop>"]
-license = "MIT"
-description = "Temporary crate used to implement the `lens!` procedural macro."
-keywords = ["plausible", "lens", "functional"]
-homepage = "https://github.com/plausiblelabs/lens-rs"
-repository = "https://github.com/plausiblelabs/lens-rs"
-readme = "README.md"
-
-[dependencies]
-pl-lens-macros-hack-impl = { version = "1.0.0", path = "../lens-macros-hack-impl" }
-proc-macro-hack = "0.5"
