+++ lens-macros/Cargo.toml
@@ -4,7 +4,7 @@ version = "1.0.0"
 edition = "2018"
 authors = ["Chris Campbell <campbell@plausible.coop>"]
 license = "MIT"
-description = "Provides procedural macros that are used in conjuction with the `pl-lens` crate."
+description = "Temporary crate used to implement the `lens!` procedural macro."
 keywords = ["plausible", "lens", "functional"]
 homepage = "https://github.com/plausiblelabs/lens-rs"
 repository = "https://github.com/plausiblelabs/lens-rs"
@@ -15,6 +15,6 @@ proc-macro = true
 
 [dependencies]
 proc-macro2 = "1.0"
+proc-macro-hack = "0.5"
 quote = "1.0"
-syn = "1.0"
-pl-lens-macros-hack-impl = { version = "1.0.0", path = "../lens-macros-hack-impl" }
+syn = { version = "1.0", features = ["full", "extra-traits"] }
