+++ lens-macros/README.md
@@ -1,7 +1,8 @@
 # pl-lens-macros
 
-Provides the custom `Lenses` derive macro that is used in conjunction with
-the `pl-lens` crate.
+This is a temporary crate used to implement the `lens!` procedural macro, which requires
+the `proc-macro-hack` crate until such a time that procedural macros are officially
+supported in expression position.
 
 This crate is automatically re-exported by the `pl-lens` crate, so you generally
 don't need to work with it directly.
